+++ Source/Core/PureES.Tests/Generators/IncrementalGeneratorTests.Generate.DotNet8_0#PureES.CommandHandlers.longArrayCommandHandler.g.verified.cs
@@ -128,7 +128,7 @@ namespace PureES.CommandHandlers
                             if (pair.Key == streamId)
                             {
                                 transaction.Add(pair.Key, new global::PureES.UncommittedEventsList(currentRevision, pair.Value));
-                                revision = currentRevision + (ulong)pair.Value.Count;
+                                revision = currentRevision + (uint)pair.Value.Count;
                             }
                             else
                             {
