+++ Source/Core/PureES/IAggregateStore.cs
@@ -4,6 +4,6 @@
 public interface IAggregateStore<T> where T : notnull
 {
     Task<T> Load(string streamId, CancellationToken cancellationToken);
-    Task<T> Load(string streamId, ulong expectedRevision, CancellationToken cancellationToken);
-    Task<T> LoadAt(string streamId, ulong endRevision, CancellationToken cancellationToken);
+    Task<T> Load(string streamId, uint expectedRevision, CancellationToken cancellationToken);
+    Task<T> LoadAt(string streamId, uint endRevision, CancellationToken cancellationToken);
 }
