+++ Source/Core/PureES/IEventEnvelope.cs
@@ -14,7 +14,7 @@ public interface IEventEnvelope : IEquatable<IEventEnvelope>
     public string StreamId { get; }
 
     /// <summary>The position of the event within the stream</summary>
-    public ulong StreamPosition { get; }
+    public uint StreamPosition { get; }
 
     /// <summary>The UTC timestamp that the event was persisted</summary>
     public DateTime Timestamp { get; }
